+++ packages/contracts/interfaces/IAntseedP0Registry.sol
@@ -1,6 +1,0 @@
-// SPDX-License-Identifier: MIT
-pragma solidity ^0.8.24;
-
-interface IAntseedP0Registry {
-    function isSellerP0(address seller) external view returns (bool);
-}